requirements.txt
streamlit
requests
python-dotenv
pandas
beautifulsoup4
